Missing video formats
Added a check whether the video returns formats that can be used to view
the video. If not, this might be due to region locks. At the moment the
code just returns. For playlists this will likely not work. So a
somewhat handling for region locked videos in playlists is still
necessary, so that the playlist plays on after skipping.